Update due to fix change/state and new lammps rods library version...
... membrane not fully functional (concentration param)

+++ programs/membrane.py
@@ -11,6 +11,10 @@ Created on 16 Mar 2018
 
 @author: Eugen Rožić
 '''
+from mpi4py import MPI
+
+mpi_comm = MPI.COMM_WORLD
+mpi_rank = mpi_comm.Get_rank()
 
 import os
 import argparse
@@ -35,8 +39,8 @@ parser.add_argument('--out', type=str, default=None,
 parser.add_argument('-o', '--output_freq', type=int,
                     help='configuration output frequency (in MD steps);'\
                     ' default behavior is after every batch of MC moves')
-parser.add_argument('-s', '--silent', action='store_true',
-                    help="doesn't print anything to stdout")
+parser.add_argument('-v', '--verbose', action='store_true',
+                    help="print all LAMMPS output to stdout")
 
 args = parser.parse_args()
 
@@ -47,9 +51,12 @@ if not args.run_file.endswith('.run'):
     raise Exception('Run configuration file (second arg) has to end with ".run"!')
 
 if args.seed is None:
-    import time
-    seed = int((time.time() % 1)*1000000)
-    print "WARNING: no seed given explicitly; using:", seed
+    seed = 0
+    if mpi_rank == 0:
+        import time
+        seed = int((time.time() % 1)*1000000)
+        print "WARNING: no seed given explicitly; using:", seed
+    seed = mpi_comm.bcast(seed, root = 0)
 else:
     seed = args.seed
 
@@ -79,7 +86,7 @@ class Membrane(object):
     main_group = 'membrane'
     
     def __init__(self, sigma, wc, eps, Nx, Ny, zmax, above, below,
-                 bead_mass=1.0, type_offset=0, bond_offset=0, angle_offset=0):
+                 bead_mass = 1.0, type_offset = 0, bond_offset = 0, angle_offset = 0):
         '''
         sigma: the characteristic length of the membrane particles (diameter of beads)
         wc: the tail-tail interaction (cosine/squared) range
@@ -126,7 +133,7 @@ class Membrane(object):
         self.angles = []
         self.angle_id = 0
     
-    def _add_lipid(self, x, y, reverse=False):
+    def _add_lipid(self, x, y, reverse = False):
         self.mol_id += 1
         for i in range(3):
             self.atom_id += 1
@@ -138,13 +145,13 @@ class Membrane(object):
                 self.atom_id, self.mol_id, self.bead_types[i], x, y, z))
         self.bond_id += 1
         self.bonds.append("\t{:d} {:d} {:d} {:d}".format(
-            self.bond_id, self.bond_type, self.atom_id-2, self.atom_id-1))
+            self.bond_id, self.bond_type, self.atom_id - 2, self.atom_id-1))
         self.bond_id += 1
         self.bonds.append("\t{:d} {:d} {:d} {:d}".format(
-            self.bond_id, self.bond_type, self.atom_id-1, self.atom_id))
+            self.bond_id, self.bond_type, self.atom_id - 1, self.atom_id))
         self.angle_id += 1
         self.angles.append("\t{:d} {:d} {:d} {:d} {:d}".format(
-            self.angle_id, self.angle_type, self.atom_id-2, self.atom_id-1, self.atom_id))
+            self.angle_id, self.angle_type, self.atom_id - 2, self.atom_id-1, self.atom_id))
         
     def generate_data(self):
         '''
@@ -191,7 +198,7 @@ class Membrane(object):
             dat_file.write("\n\nAngles\n\n")
             dat_file.write('\n'.join(self.angles))
             
-    def create_membrane(self, py_lmp, seed, append=True):
+    def create_membrane(self, py_lmp, seed, append = True):
         '''
         Creates the membrane using the "read_data" command. If not already done, the
         membrane data for this object will be generated.
@@ -204,7 +211,8 @@ class Membrane(object):
         if self.atom_id == 0:
             self.generate_data()
         temp_dat_file = str(seed)+'-temp-membrane.dat'
-        self.output_data(temp_dat_file)
+        if mpi_rank == 0:
+            self.output_data(temp_dat_file)
         
         all_atom_mol_ids = py_lmp.lmp.gather_atoms_concat("molecule", 0, 1)
         if len(all_atom_mol_ids) > 0:
@@ -217,7 +225,7 @@ class Membrane(object):
             read_data_args.append("add append")
 #             read_data_args.append("offset {:d} {:d} {:d} 0 0".format(
 #                 self.type_offset, self.bond_offset, self.angle_offset))
-        py_lmp.read_data('"'+temp_dat_file+'"', ' '.join(read_data_args))
+        py_lmp.read_data('"' + temp_dat_file + '"', ' '.join(read_data_args))
         
         py_lmp.group(Membrane.top_layer_group, "molecule",
                      "{:d}:{:d}:2".format(mol_offset+1, mol_offset+membrane.mol_id))
@@ -225,15 +233,14 @@ class Membrane(object):
                      "{:d}:{:d}:2".format(mol_offset+2, mol_offset+membrane.mol_id))
         py_lmp.group(Membrane.main_group, "type", *membrane.bead_types)
         
-        os.remove(temp_dat_file)
+        if mpi_rank == 0:
+            os.remove(temp_dat_file)
 
 # =======================================================================================
-
-#from mpi4py import MPI #TODO make MPI work...
 from lammps import PyLammps
 import lammps_multistate_rods as rods
 
-if not os.path.exists(output_folder):
+if not os.path.exists(output_folder) and mpi_rank == 0:
     os.makedirs(output_folder)
     
 run_filename = os.path.splitext(os.path.basename(args.run_file))[0]
@@ -245,22 +252,32 @@ dump_path = os.path.join(output_folder, dump_filename)
 log_filename = '{:d}.lammps'.format(seed)
 log_path = os.path.join(output_folder, log_filename)
 
-run_args = rods.rod_model.Params()
-execfile(args.run_file,
-         {'__builtins__' : None, 'True' : True, 'False' : False, 'None' : None},
-         vars(run_args))
+run_args = rods.Rod_params() # any object with __dict__ would do
+if mpi_rank == 0:
+    execfile(args.run_file,
+             {'__builtins__' : None, 'True' : True, 'False' : False, 'None' : None},
+             vars(run_args))
+run_args = mpi_comm.bcast(run_args, root = 0)
 
-out_freq = args.output_freq if args.output_freq != None else run_args.run_length
+out_freq = args.output_freq if args.output_freq != None else run_args.mc_every
 
-model = rods.Rod_model(args.cfg_file)
+rod_params = rods.Rod_params()
+if mpi_rank == 0:
+    rod_params.from_file(args.cfg_file);
+rod_params = mpi_comm.bcast(rod_params, root = 0)
+
+model = rods.Rod_model(rod_params)
 membrane = Membrane(run_args.mem_sigma, run_args.mem_wc, run_args.mem_eps, 
                     run_args.mem_Nx, run_args.mem_Ny, 0.0, model.rod_length/2, 5.0)
 
 # ===== LAMMPS setup ====================================================================
-py_lmp = PyLammps(cmdargs=['-screen','none'])
-py_lmp.log('"'+log_path+'"')
+if mpi_rank == 0:
+    py_lmp = PyLammps(cmdargs = ['-echo', 'both'], comm = mpi_comm, verbose = args.verbose)
+else:
+    py_lmp = PyLammps(cmdargs = ['-echo', 'both'], comm = mpi_comm)
+py_lmp.log('"' + log_path + '"')
 
-simulation = rods.Simulation(py_lmp, model, seed, output_folder)
+simulation = rods.Simulation(py_lmp, model, run_args.temp, seed, output_folder)
 
 py_lmp.units('lj')
 py_lmp.dimension(3)
@@ -270,35 +287,35 @@ py_lmp.region('box', 'block',
               membrane.ymin, membrane.ymax,
               0.0, run_args.Lz)
 
-simulation.setup('box', type_offset=max(membrane.bead_types),
-                 extra_pair_styles=[('cosine/squared', model.global_cutoff)],
-                 bond_offset=membrane.bond_type, extra_bond_styles=['fene'],
-                 everything_else=['angle/types', 1, 'extra/angle/per/atom', 1])
+simulation.setup('box', type_offset = max(membrane.bead_types),
+                 extra_pair_styles = [('cosine/squared', model.global_cutoff)],
+                 bond_offset = membrane.bond_type, extra_bond_styles=['fene'],
+                 opt = ['angle/types', 1, 'extra/angle/per/atom', 1])
 py_lmp.angle_style('harmonic')
 
 zwalls_fix = 'zwalls'
 py_lmp.fix(zwalls_fix, 'all', 'wall/lj126',
-           'zlo EDGE', 1.0, model.rod_radius, model.rod_radius*pow(2,1./6),
-           'zhi EDGE', 1.0, model.rod_radius, model.rod_radius*pow(2,1./6))
+           'zlo EDGE', 1.0, model.rod_radius, model.rod_radius * pow(2, 1./6),
+           'zhi EDGE', 1.0, model.rod_radius, model.rod_radius * pow(2, 1./6))
 
 for bead_type in membrane.bead_types:
     py_lmp.mass(bead_type, membrane.bead_mass)
-py_lmp.angle_coeff(membrane.angle_type, 5.0*membrane.eps, 180)
-py_lmp.bond_coeff(membrane.bond_type, 'fene', 30.0*membrane.eps/membrane.sigma**2,
-                  1.5*membrane.sigma, membrane.eps, membrane.sigma)
+py_lmp.angle_coeff(membrane.angle_type, 5.0 * membrane.eps, 180)
+py_lmp.bond_coeff(membrane.bond_type, 'fene', 30.0 * membrane.eps / membrane.sigma**2,
+                  1.5 * membrane.sigma, membrane.eps, membrane.sigma)
 
 lj_factor = pow(2, 1./6)
 # head-head & head-tail inter-lipid interaction
 for bead_type in membrane.bead_types:
     py_lmp.pair_coeff(membrane.head_type, bead_type, membrane.eps,
-                      0.95*membrane.sigma*lj_factor,
-                      0.95*membrane.sigma*lj_factor, 'wca')
+                      0.95 * membrane.sigma * lj_factor,
+                      0.95 * membrane.sigma * lj_factor, 'wca')
 # tail-tail inter-lipid interaction
 for i in range(len(membrane.tail_types)):
     for j in range(i, len(membrane.tail_types)):
         py_lmp.pair_coeff(membrane.tail_types[i], membrane.tail_types[j], membrane.eps,
-                          1.0*membrane.sigma*lj_factor,
-                          (1.0*lj_factor + membrane.wc)*membrane.sigma, 'wca')
+                          1.0 * membrane.sigma * lj_factor,
+                          (1.0 * lj_factor + membrane.wc) * membrane.sigma, 'wca')
 
 # lipid-protein interaction
 sol_lipid_eps = run_args.mem_int_eps
@@ -308,19 +325,19 @@ rod_type_range = "{:d}*{:d}".format(simulation.get_min_rod_type(),
                                     simulation.get_max_rod_type())
 mem_type_range = "{:d}*{:d}".format(min(membrane.bead_types),
                                     max(membrane.bead_types))
-py_lmp.pair_coeff(mem_type_range, rod_type_range, 0.0, model.global_cutoff-0.01, model.global_cutoff)
+py_lmp.pair_coeff(mem_type_range, rod_type_range, 0.0, model.global_cutoff - 0.01, model.global_cutoff)
 # lipid-protein volume-exclusion (all states)
 vx_types = filter(lambda t: t not in model.active_bead_types,
                   model.all_bead_types)
 for vx_type in vx_types:
-    bead_lipid_contact = 0.5*membrane.sigma*pow(2,1./6) + model.bead_radii[vx_type]
+    bead_lipid_contact = 0.5 * membrane.sigma * pow(2, 1./6) + model.bead_radii[vx_type]
     for mem_bead_type in membrane.bead_types:
         py_lmp.pair_coeff(mem_bead_type, vx_type + simulation.type_offset,
                           sol_lipid_eps, bead_lipid_contact, bead_lipid_contact, 'wca')
 # lipid-protein tip interaction (all states)
 tip_types = [state_struct[0][-1] + simulation.type_offset
              for state_struct in model.state_structures]
-tip_lipid_contact = 0.5*membrane.sigma*pow(2,1./6) + model.rod_radius
+tip_lipid_contact = 0.5 * membrane.sigma * pow(2, 1./6) + model.rod_radius
 tip_lipid_cutoff = tip_lipid_contact + run_args.mem_int_range
 for tip_type in tip_types:
     for mem_bead_type, k in zip(membrane.bead_types, int_factors):
@@ -349,20 +366,16 @@ if hasattr(run_args, 'label_micelles'):
 # FIXES & DYNAMICS
 thermo_fix = 'thermostat'
 py_lmp.fix(thermo_fix, 'all', 'langevin', run_args.temp, run_args.temp, run_args.damp, seed)#, "zero yes")
-simulation.set_rod_dynamics('nve', everything_else=['mol', model.rod_states[0]])
+
+simulation.set_rod_dynamics('nve', opt = ['mol', model.rod_states[0]])
 
 py_lmp.region('gcmc_init', 'block',
               membrane.xmin + model.rod_length/2, membrane.xmax - model.rod_length/2,
               membrane.ymin + model.rod_length/2, membrane.ymax - model.rod_length/2,
               0.0 + model.rod_length/2, run_args.Lz - model.rod_length/2)
-rod_gcmc_fix = 'rod_gcmc'
-py_lmp.fix(rod_gcmc_fix, simulation.rods_group, 'gcmc', 100, 200, 0,
-           0, seed, run_args.temp, run_args.mu, 0.0,
-           'region', 'gcmc_init', 'mol', model.rod_states[0],
-           'rigid', simulation.rod_dyn_fix, 'tfac_insert', 1.65)
-
-#py_lmp.fix_modify(simulation.rod_dyn_fix, 'dynamic/dof yes') #only for nvt&npt (small)
-py_lmp.compute_modify("thermo_temp", "dynamic/dof yes")
+simulation.set_state_concentration(0, run_args.conc, 100, 200,
+                                   opt = ['region', 'gcmc_init',
+                                          'tfac_insert', 1.65])
 
 # TEST DUMP...
 # py_lmp.thermo_style('custom', 'step atoms', 'pe temp')
@@ -375,15 +388,15 @@ py_lmp.compute_modify("thermo_temp", "dynamic/dof yes")
 # GENERATING INITIAL CONFIGURATION
 py_lmp.neigh_modify('every', 1, 'delay', 1)
 py_lmp.timestep(run_args.dt)
-py_lmp.command('run 2000')
-py_lmp.unfix(rod_gcmc_fix)
+py_lmp.run(2000)
+simulation.unset_state_concentration(0)
 py_lmp.unfix(zwalls_fix)
 py_lmp.reset_timestep(0)
 
 # ===== MEMBRANE ========================================================================
 
 # create membrane (box update, create membrane & groups, ...)
-membrane.create_membrane(py_lmp, seed, append=True)
+membrane.create_membrane(py_lmp, seed, append = True)
 
 py_lmp.fix(zwalls_fix, 'all', 'wall/lj126',
            'zlo EDGE', 1.0, model.rod_radius, model.rod_radius*pow(2,1./6),
@@ -408,57 +421,58 @@ py_lmp.compute(full_msd_compute, membrane.main_group, 'msd')
 
 # FIXES & DYNAMICS
 mem_dyn_fix = 'mem_dynamics'
-press_eq_t = 1000*run_args.dt
+press_eq_t = 1000 * run_args.dt
 py_lmp.fix(mem_dyn_fix, membrane.main_group, 'nph',
            'x 0.0 0.0', press_eq_t, 'y 0.0 0.0', press_eq_t, 'couple xy',
            'dilate', 'all')
-py_lmp.compute_modify(mem_dyn_fix+'_temp', 'dynamic/dof yes')
+py_lmp.compute_modify(mem_dyn_fix + '_temp', 'dynamic/dof yes')
 py_lmp.neigh_modify('exclude', 'molecule/intra', membrane.main_group)
 # -> saves neighbour calculating without consequences (because beads 1 & 3 will
 #    never be in range anyway & this doesn't affect bonds and angles)
 py_lmp.special_bonds('fene') # not really necessary because of "neigh_modify exclude"
 
 # MEMBRANE EQUILIBRATION
-py_lmp.command('run 2000') #2 x press_eq_t, for good measure
+py_lmp.run(2000) #2 x press_eq_t, for good measure
 py_lmp.reset_timestep(0)
 
 # ===== FINAL ===========================================================================
-py_lmp.variable('temp_x', 'equal', '"xhi - {:f}"'.format(model.rod_length/2))
-py_lmp.variable('temp_y', 'equal', '"yhi - {:f}"'.format(model.rod_length/2))
+py_lmp.variable('temp_x', 'equal', '"xhi - {:f}"'.format(model.rod_length / 2))
+py_lmp.variable('temp_y', 'equal', '"yhi - {:f}"'.format(model.rod_length / 2))
 py_lmp.region('gcmc_box', 'block',
               '-${temp_x}', '${temp_x}',
               '-${temp_y}', '${temp_y}',
-              0.0 + 2*model.rod_length, run_args.Lz - model.rod_length/2)
-py_lmp.fix(rod_gcmc_fix, simulation.rods_group, 'gcmc', 1000, 10, 0,
-           0, seed, run_args.temp, run_args.mu, 0.0,
-           'region', 'gcmc_box', 'mol', model.rod_states[0],
-           'rigid', simulation.rod_dyn_fix, 'tfac_insert', 1.65)
+              0.0 + 2 * model.rod_length, run_args.Lz - model.rod_length / 2)
 
+simulation.set_state_concentration(0, run_args.conc, 1000, 10,
+                                   opt = ['region', 'gcmc_box',
+                                          'tfac_insert', 1.65])
 # OUTPUT
-py_lmp.variable("area", "equal", "lx*ly")
-py_lmp.thermo_style("custom", "step atoms", "pe temp", "press lx ly v_area")
-py_lmp.thermo(out_freq)
 dump_elems = "id x y z type mol"
 try:
-    dump_elems += " c_"+micelle_compute
+    dump_elems += " c_" + micelle_compute
 except:
     pass
 try:
-    dump_elems += " c_"+adsorbed_compute
+    dump_elems += " c_" + adsorbed_compute
 except:
     pass
 py_lmp.dump("dump_cmd", "all", "custom", out_freq, dump_path, dump_elems)
 py_lmp.dump_modify("dump_cmd", "sort id")
+py_lmp.variable("area", "equal", "lx*ly")
+py_lmp.thermo_style("custom", "step atoms", "pe temp", "press lx ly v_area")
+py_lmp.thermo(out_freq)
 # MSD data (only in x & y directions)
 py_lmp.variable("top_msd", "equal", '"c_{0}[1] + c_{0}[2]"'.format(top_msd_compute))
 py_lmp.fix("mem_top_msd", "all", "ave/time", 1, 1, out_freq, "v_top_msd",
-           "file", os.path.join(output_folder, sim_ID+'_mem_top.msd'))
+           "file", os.path.join(output_folder, sim_ID + '_mem_top.msd'))
 py_lmp.variable("bottom_msd", "equal", '"c_{0}[1] + c_{0}[2]"'.format(bottom_msd_compute))
 py_lmp.fix("mem_bottom_msd", "all", "ave/time", 1, 1, out_freq, "v_bottom_msd",
-           "file", os.path.join(output_folder, sim_ID+'_mem_bottom.msd'))
+           "file", os.path.join(output_folder, sim_ID + '_mem_bottom.msd'))
 py_lmp.variable("full_msd", "equal", '"c_{0}[1] + c_{0}[2]"'.format(full_msd_compute))
 py_lmp.fix("mem_full_msd", "all", "ave/time", 1, 1, out_freq, "v_full_msd",
-           "file", os.path.join(output_folder, sim_ID+'_mem_full.msd'))
+           "file", os.path.join(output_folder, sim_ID + '_mem_full.msd'))
 
 # RUN...
-py_lmp.command('run {:d}'.format(args.simlen))
+py_lmp.run(args.simlen)
+
+MPI.Finalize()
